Rename single and multiple tasks (#4102)
All tasks now live in directories under app/classifier/tasks/.
